+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Make identity pool id header optional, with union-of-pools support (#30)
 * Add support for saving Azure key version with DEK (#31)
+* Pass context when clients make KEK calls to DEK Registry (#32)
 
 ## Fixes